Se realizan endpoint para registrar usuarios y enspoint para la autenticación
