+++ DesafioFinalV1.0/src/main/resources/application.properties
@@ -1,3 +1,4 @@
+jwt.secret = team1bootcamp
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.url=jdbc:h2:mem:testdb
