--- DesafioFinalV1.0/src/test/java/com/example/demo2/Unit/Demo2ApplicationTests.java
@@ -1,24 +1,0 @@
-package com.example.demo2.Unit;
-
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.springframework.boot.test.context.SpringBootTest;
-
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.*;
-import java.util.Arrays;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.when;
-
-
-@SpringBootTest
-class Demo2ApplicationTests {
-	@Test
-	void contextLoads() {
-	}
-}
