--- DesafioFinalV1.0/src/test/java/com/example/demo2/Unit/utils/JwtTokenUtilTest.java
@@ -1,14 +1,0 @@
-package com.example.demo2.Unit.utils;
-
-import com.example.demo_bootcamp_spring.util.JwtTokenUtil;
-import org.mockito.InjectMocks;
-import org.springframework.boot.test.context.SpringBootTest;
-
-
-@SpringBootTest
-class JwtTokenUtilTest {
-
-    @InjectMocks
-    private JwtTokenUtil jwtTokenUtil;
-
-}
