+++ DesafioFinalV1.0/src/test/java/com/example/demo_bootcamp_spring/Demo2ApplicationTests.java
@@ -1,0 +1,10 @@
+package com.example.demo_bootcamp_spring;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Demo2ApplicationTests {
+	@Test
+	void contextLoads() {
+	}
+}
