+++ DesafioFinalV1.0/src/test/java/com/example/demo_bootcamp_spring/utils/JwtTokenUtilTest.java
@@ -1,0 +1,14 @@
+package com.example.demo_bootcamp_spring.utils;
+
+import com.example.demo_bootcamp_spring.util.JwtTokenUtil;
+import org.mockito.InjectMocks;
+import org.springframework.boot.test.context.SpringBootTest;
+
+
+@SpringBootTest
+class JwtTokenUtilTest {
+
+    @InjectMocks
+    private JwtTokenUtil jwtTokenUtil;
+
+}
